Fixed location of manage.py command + load code from filename
